Backend/hackathon/requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
pypdf
supabase
python-dotenv
tiktoken